Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/webmessaging/without-ports/028-expected.txt


FAIL Cloning objects, preserving sharing #2 assert_equals: expected object "[object ImageData]" but got object "[object ImageData]"
